Clean up Makefile
Removed the Darwin and Windows specific build targets.  Instead, the
system's build target will be used (as specified with the GOOS and
GOARCH environment variables).

Added a "server" build target.  This is mostly just for me to build the
binary that will be deployed on my own server.

Added the GO_VERSION environment variable to all the calls to the go
command.  This will allow easier building on systems that have installed
an additional version of Go along side their distribution's version.
Users wanting to build with one of these versions must first install it
with `go get` and `go[version] download` as detailed here:
https://golang.org/doc/install#extra_versions.

Setting the GO_VERSION environment variable to the version number will
use that version. Eg, setting it to "1.13.7" will call `go1.13.7 build
[...]` instead of `go build [...]`.