Add unlocks to error cases in handlers::handlePublish()
In handlers::handlePublish(), the mutex for the rtmp stream is locked at the beginning of the function. The previous implementation checked for error cases (e.g. incorrect stream key), logging and immediately returning if an error case was found. The mutex, however, was not unlocked, leading to the server effectively 'freezing' as other interactions try and lock that mutex. 
This MR adds mutex unlocks in each error case to ensure that the server continues to function in the event of a recoverable error.